If you don't give ids on the command-line to rocket_control_soap, it doesn't barf anymore; rather, it'll wiggle each rocket and request ids for them.
The Rocket class now accepts a default (or delta) heading along with location to determine where a "center" method leaves the rocket pointing.

The RocketArray class has a slightly shortened rollcall(), and a slightly lenghtened (and refined) cancan()

The RocketTurtle class now turns when it's supposed to and recenters to the default heading.